test(nexus): refactoring to integration test
