+++ crates/nexus/src/command/mod.rs
@@ -24,64 +24,9 @@
 //! auditable systems where state mutations are explicit, event-driven, and
 //! managed through well-defined domain models.
 //!
-use crate::domain::DomainEvent;
-use smallvec::{SmallVec, smallvec};
 
 pub mod handler;
 pub mod repository;
 
-#[cfg(test)]
-pub mod test;
-
-#[derive(Debug)]
-pub struct Events<E>
-where
-    E: DomainEvent,
-{
-    first: E,
-    more: SmallVec<[E; 1]>,
-}
-
-impl<E> Events<E>
-where
-    E: DomainEvent,
-{
-    pub fn new(event: E) -> Self {
-        Events {
-            first: event,
-            more: SmallVec::new(),
-        }
-    }
-
-    pub fn add(&mut self, event: E) {
-        self.more.push(event);
-    }
-
-    pub fn into_small_vec(self) -> SmallVec<[E; 1]> {
-        let mut events = smallvec![self.first];
-        events.extend(self.more);
-        events
-    }
-}
-
-// TODO: impl IntoIterator for this collection
-// TODO: impl From trait to small_vec
-
-#[macro_export]
-macro_rules! events {
-    [$head:expr] => {
-        {
-            let mut events = Events::new($head);
-            events
-        }
-    };
-    [$head:expr, $($tail:expr), +] => {
-        {
-            let mut events = Events::new($head);
-            $(
-                events.add($tail);
-            )*
-                events
-        }
-    }
-}
+pub use handler::{AggregateCommandHandler, CommandHandlerResponse};
+pub use repository::{EventSourceRepository, RepositoryError};
